MyWeb/requirements.txt
allure-pytest==2.14.0
allure-python-commons==2.14.0
asgiref==3.8.1
attrs==25.1.0
bleach==6.2.0
blinker==1.9.0
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.2
cssselect==1.3.0
cycler==0.12.1
Django==5.2.1
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-markdownx==4.0.9
docopt==0.6.2
easytrader==0.23.6
easyutils==0.1.7
execnet==2.1.1
Flask==3.1.1
fonttools==4.58.0
h11==0.14.0
idna==3.10
iniconfig==2.1.0
ipython==8.12.3
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mistune==3.1.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
nbconvert==7.16.6
numpy==2.2.5
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pipreqs==0.5.0
pluggy==1.5.0
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.2.3
pyquery==2.0.1
PySocks==1.7.1
pytest==8.3.5
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.0
pytest-ordering==0.6
pytest-rerunfailures==15.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywinauto==0.6.6
requests==2.32.3
schedule==1.2.2
selenium==4.28.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
tushare==1.4.21
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarg==0.1.9
